pyproject.toml
[project]
name = "exis-algorithm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "grandalf>=0.8",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-cli[inmem]>=0.4.0",
    "python-dotenv>=1.1.1",
    "fastapi==0.118.3",
    "openai-agents==0.3.3",
    "uvicorn==0.37.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
